+++ src/app/yoga-day-26/_components/data.ts
@@ -1,0 +1,77 @@
+// ─── Static Data ──────────────────────────────────────────────────────────────
+
+export const SCHEDULE = [
+    { time: '6:00 am – 6:30 am', activity: 'Registration & Welcome', highlight: false },
+    { time: '6:30 am – 6:45 am', activity: 'Arrival of Dignitaries', highlight: false },
+    { time: '6:45 am – 7:00 am', activity: 'Inauguration Ceremony', highlight: true },
+    { time: '7:00 am – 8:00 am', activity: 'Mass Yoga Session', note: 'Led by Athayog Team', highlight: true },
+    { time: '8:00 am – 8:05 am', activity: 'Talent Showcase', highlight: false },
+    { time: '8:05 am – 8:10 am', activity: 'Address by Founder, Athayog', highlight: false },
+    { time: '8:10 am – 8:30 am', activity: 'Address by Dignitaries', highlight: false },
+    { time: '8:30 am – 8:35 am', activity: 'Valedictory & Closing Ceremony', highlight: false },
+    { time: '8:35 am onwards', activity: 'Healthy Refreshments & Networking', highlight: false },
+]
+
+export const WHY_ITEMS = [
+    {
+        title: 'Wellness for Women, Wellness for the World',
+        body: 'When a woman thrives, communities flourish. Yoga nurtures mind, body, and soul — creating a ripple effect of strength and compassion that extends to families, societies, and the Earth itself.',
+    },
+    {
+        title: 'Yoga for Every Body, Every Being',
+        body: 'Yoga knows no boundaries — welcoming people of all ages, backgrounds, and walks of life. When we come together on the mat, we heal ourselves and contribute to collective planetary health.',
+    },
+    {
+        title: 'Inner Calm, Outer Care',
+        body: "Yoga is more than movement — it's mindfulness that grounds us in the present and awakens responsibility to nature. Through breath, balance, and awareness, we create harmony within and with the Earth.",
+    },
+]
+
+export const EXPERIENCE_ITEMS = [
+    { num: '01', title: 'Yoga Practice', body: 'Tailored to all levels, focusing on postures and flows that promote strength, flexibility and energy.' },
+    { num: '02', title: 'Breath Work', body: 'Guided practices to calm the mind, reduce stress and enhance vitality.' },
+    { num: '03', title: 'Community Connection', body: 'Engage with a vibrant community in a warm, supportive and uplifting environment.' },
+    { num: '04', title: 'Enriching Experiences', body: 'Discover holistic benefits of yoga through experiences that go beyond the mat.' },
+]
+
+export const ACTIVITIES = [
+    { title: 'Live Music', body: 'Soothing tunes elevating the yoga atmosphere' },
+    { title: 'Live Painting', body: 'Yoga-inspired art created live on canvas' },
+    { title: 'AI Station', body: 'Where technology meets wellness' },
+    { title: 'Blood Donation', body: 'Give back through our wellness drive' },
+    { title: 'Yoga as Career', body: 'Turn passion into profession' },
+    { title: 'Yoga Clinics', body: 'Therapy, rehab & holistic health' },
+]
+
+export const DIGNITARIES = [
+    { tag: 'Presided by', name: 'Shri P.C. Mohan', role: 'MP, Central Bengaluru' },
+    { tag: 'Chief Guest', name: 'Shri Tejasvi Surya', role: 'MP, South Bengaluru' },
+    { tag: 'Guest of Honour', name: 'Shri BNS Reddy', role: 'Indiranagar Club President, Ex IPS' },
+]
+
+export const FAQS = [
+    {
+        q: 'Is Yoga Arambha 2026 free to attend?',
+        a: 'Yes, completely free and open to everyone. No yoga experience required. Simply register and show up.',
+    },
+    {
+        q: 'Where is the event held?',
+        a: 'Indiranagar Club, Bangalore on Sunday, June 21, 2026.',
+    },
+    {
+        q: 'Do I need prior yoga experience?',
+        a: 'Not at all. Sessions are tailored to every level — from complete beginners to experienced practitioners.',
+    },
+    {
+        q: 'What time should I arrive?',
+        a: 'On-site registration and welcome begins at 6:00 AM. Dignitaries arrive at 6:30 AM and the mass yoga session runs from 7:00–8:00 AM.',
+    },
+    {
+        q: 'Who are the dignitaries at Yoga Arambha 2026?',
+        a: 'The event is presided by Shri P.C. Mohan (MP, Central Bengaluru). Chief Guest is Shri Tejasvi Surya (MP, South Bengaluru). Guest of Honour is Shri BNS Reddy (Indiranagar Club President, Ex IPS).',
+    },
+    {
+        q: 'What is the theme of Yoga Arambha 2026?',
+        a: '"Yoga for Wellness, Wisdom & World Peace" — celebrating yoga as a tool for holistic health, mental well-being, and social harmony.',
+    },
+]
